✨ feat: shopMypage api 개발

+++ src/main/java/org/example/flow/repository/ShopInfoRepository.java
@@ -1,8 +1,10 @@
 package org.example.flow.repository;
 
-import org.example.flow.entity.ShopInfo;
+import org.example.flow.entity.*;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.repository.CrudRepository;
 
-public interface ShopInfoRepository extends JpaRepository<ShopInfo, Integer> {
-}
+import java.util.List;
+import java.util.Optional;
+
+public interface ShopInfoRepository extends JpaRepository<ShopInfo, Long> { }
+
